requirements.txt
# Core Flask dependencies
Flask==2.3.3
gunicorn==21.2.0
python-dotenv==1.0.0

# Security
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
bcrypt==4.0.1
bleach==6.0.0
cryptography==41.0.4
python-clamd==0.4.0  # ClamAV integration for virus scanning (HIPAA required)

# Database (supports both SQLite and PostgreSQL)
# Using psycopg (v3) for Python 3.13 compatibility
psycopg[binary]==3.2.3

# Document Processing (local, no cloud costs except Azure OpenAI)
PyPDF2==3.0.1           # PDF parsing
python-docx==1.1.0      # Word doc parsing
Pillow==10.4.0          # Image processing (compatible with Python 3.13)
pytesseract==0.3.10     # OCR (free, uses Tesseract)

# Azure OpenAI (ONLY RECURRING COST - for intelligent extraction)
openai>=1.3.0

# Data processing
pandas==2.2.3           # CSV handling and data manipulation

# Utilities
Werkzeug==2.3.7
requests==2.31.0

# Production Infrastructure
boto3==1.28.0  # AWS S3 for file storage (optional)
azure-storage-blob==12.19.0  # Azure Blob Storage (optional)
celery==5.3.4  # Background task processing
redis==5.0.1  # Task queue backend
sentry-sdk[flask]==1.39.0  # Error tracking
